Migrated Tests.csproj to the sdk-based CSProj, added multi-targeted options to test across the whole spectrum
Because test discovery and execution works a little different in modern visual studio, we have to remove the old strong dependencies onto NUnit.
The test project also doesn't have to be an executable anymore - with the new TestAdapter system we
can just add the NUnit and NUnit3TestAdapter packagereference and the dotnet-sdk/cli and visual have everything they need.
For all this to work together a little more nicely i also migrated the solution file to the VS2017 schema.