Migrated DataTableProxy.csproj to the sdk-based CSProj, added multi-targeted options to allow for usage in netstandard2.0 and .NET Full Framework 4.0+
With the dotnet cli/sdk installed you don't need the nuspec anymore. everything works inside the CSProj.
You just "dotnet pack" and the package builds with dependencygroups for all target frameworks listed in the <targetframeworks> property.

NuGet depecrated the owners attribute, which is why it did not move into the csproj everything else is pretty copypasta'd from the nuspec into the csproj.
The AssemblyInfo is removed because sdk-based csproj builds autogenerate these based on the Version information inside
the csproj (which is better because in our example we already see the issues of multiple sources of truth; the AssemblyInfo
says version 1.1.0.0 and the nuspec says 1.2).